Renamed components from Number to Tab
